+++ src/main/java/com/traveloveapi/entity/service_package/bill/voucher/BillVoucherEntity.java
@@ -4,7 +4,6 @@ import jakarta.persistence.Entity;
 import jakarta.persistence.Id;
 import jakarta.persistence.IdClass;
 import jakarta.persistence.Table;
-import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Entity
@@ -18,9 +17,9 @@ public class BillVoucherEntity {
     @Id
     private String voucher_id;
 
-    private float discount_amount;
+    private Double discount_amount;
 
-    public BillVoucherEntity(String bill_id, String voucher_id, float discount_amount) {
+    public BillVoucherEntity(String bill_id, String voucher_id, Double discount_amount) {
         this.bill_id = bill_id;
         this.voucher_id = voucher_id;
         this.discount_amount = discount_amount;
